MSn parser now stores the isolation window. Selected ions now store either a set of possible charges and the precursor m/z as the selected m/z OR a single charge, accurate mass, and selected m/z computed from the accurate mass.  The second case is for files that have been processed by Bullseye.
Former-commit-id: fcc769aea062a2fd8e769653febcdc817d9af7f9

+++ pwiz/data/msdata/MSDataMergerTest.cpp
@@ -38,7 +38,7 @@ void test()
     MSData tinyReference;
     examples::initializeTiny(tinyReference);
 
-    const int tinyCopyCount = 3;
+    const size_t tinyCopyCount = 3;
 
     vector<MSDataPtr> tinyExamples;
     for (size_t i=0; i < tinyCopyCount; ++i)
